Update 05/21/2025
I have made quite a bit of progress. I finally got pieces to fall all the way down, and i even have some very basic collision and rotation going on.

I am now working on the actual deletion of the pieces whenever you get a row cleared! This should be fully working in the next update! then i will work on scoring system, level system, and a piece preview. After that, i will work out some more of the game-breaking bugs.

Some bugs..
The 'L' and 'J' pieces are bugged with their collisions on the hook parts when they are turned upside down .This is something i plan on fixing in the future

If you rotate a piece next to another piece, it will clip into it

If you rotate a piece near the bottom of the board, it can crash the game if the piece clips the bottom of the board. This will be fixed soon

Thanks for reading
-Hunter